+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/EpochSpec.hs
@@ -88,7 +88,7 @@ proposalsSpec =
           & ppGovActionDepositL .~ deposit
       rewardAccount <- registerRewardAccount
 
-      initialValue <- getsNES (nesEsL . curPParamsEpochStateL . ppMinFeeAL)
+      initialValue <- getsNES (nesEsL . curPParamsEpochStateL . ppMinFeeFactorL)
 
       parameterChangeAction <- mkMinFeeUpdateGovAction SNothing
       govActionId <-
@@ -100,7 +100,7 @@ proposalsSpec =
       passNEpochs 3
       expectMissingGovActionId govActionId
 
-      getsNES (nesEsL . curPParamsEpochStateL . ppMinFeeAL) `shouldReturn` initialValue
+      getsNES (nesEsL . curPParamsEpochStateL . ppMinFeeFactorL) `shouldReturn` initialValue
       getAccountBalance rewardAccount `shouldReturn` deposit
 
     it "Proposals are expired and removed as expected" $ whenPostBootstrap $ do
@@ -352,11 +352,11 @@ dRepVotingSpec =
     -- so we can only run this test post-bootstrap
     it "proposal is accepted after two epochs" $ whenPostBootstrap $ do
       modifyPParams $ ppDRepVotingThresholdsL . dvtPPEconomicGroupL .~ 1 %! 1
-      let getParamValue = getsNES (nesEsL . curPParamsEpochStateL . ppMinFeeAL)
+      let getParamValue = getsNES (nesEsL . curPParamsEpochStateL . ppMinFeeFactorL)
       initialParamValue <- getParamValue
 
       let proposedValue = unCoinPerByte initialParamValue <+> Coin 300
-      let proposedUpdate = def & ppuMinFeeAL .~ SJust (CoinPerByte proposedValue)
+      let proposedUpdate = def & ppuMinFeeFactorL .~ SJust (CoinPerByte proposedValue)
 
       -- Submit NewConstitution proposal two epoch too early to check that the action
       -- doesn't expire prematurely (ppGovActionLifetimeL is set to two epochs)